refactor(canvas): rename canvas2d-state.ts -> mixins.ts and helpers.ts -> utilities.ts
Pure renames with specifier updates only (mixin and helper bodies unchanged).
canvas2DStateMixin now lives in mixins.ts; the canvas-coupled helpers move to
utilities.ts for consistent naming.
