+++ packages/canvas/src/index.ts
@@ -1,4 +1,4 @@
-export * from './canvas2d-state';
+export * from './mixins';
 export * from './context';
-export * from './helpers';
+export * from './utilities';
 export * from './path';
