+++ packages/canvas/src/utilities.ts
